reactor: move user and cpu specific functions to a new players file
